Phase 2: move autodiff.py/harrison_tb.py/vhd_tb.py into dense_evolution.solvers/ (final subpackage)
Two prog.txt corrections made discovering the real file layout, same
category as earlier _logging.py/qiskit_pennylane.py mismatches:
- "vqe.py" from the original target tree doesn't exist as a separate
  file -- VQE is circuit_to_energy_fn, already living inside autodiff.py.
- vhd_tb.py (sp3s* tight-binding, Vogl-Hjalmarson-Dow material-specific
  parameters) is a direct sibling of harrison_tb.py (same "tight-binding
  solver" category, sp3 universal parameters) and was missing from the
  original solvers/ list -- added it here.

- git mv all 3 modules into dense_evolution/solvers/
- new solvers/__init__.py re-exports the public API
- backward-compat shims at all 3 old top-level paths (checked first for
  a same-basename-as-subpackage collision like mitigation/'s mistake --
  none of autodiff/harrison_tb/vhd_tb match "solvers", so all 3 correctly
  need real shim files)
- fixed autodiff.py's 5 relative imports into circuits/ (.circuits.<mod>
  -> ..circuits.<mod>, one level deeper now)

Fixed one real regression, same private-helper/monkeypatch-target pattern
already seen in backends/: tests/test_autodiff.py did `from dense_evolution
import autodiff` then monkeypatched HAS_JAX on it -- the real check lives
in the real submodule's own globals, so pointed at
`from dense_evolution.solvers import autodiff` directly instead.

Proactively checked shim test coverage per the discipline from PR #83/#84:
dense_evolution.autodiff needed a new identity-check test (nothing else
exercises that flat path after the monkeypatch-fix redirect above, same
gap category as backends/mps.py). dense_evolution.harrison_tb and
dense_evolution.vhd_tb are already naturally exercised (100% coverage
confirmed) by their own test files' direct shim-path imports -- no fix
needed there.

Full suite: 802 passed, 17 skipped, 3 failed on first run -- all 3
MemoryPressureError. 2 of the 3 still failed on a small isolated re-run,
but both passed cleanly when run completely alone -- consistent with the
known resource-contention pattern (SafeMemoryGuard reading actual system
RAM at test time) from a full day of consecutive heavy pytest runs on
this machine, not a regression from this move.

This completes Phase 2 of Sezione 1 entirely: all 7 subpackages
(utils, interop, physics, circuits, backends, mitigation, solvers) are
now split out of the flat dense_evolution/ root, each with a CI-gated PR
and a verified backward-compat shim (or unchanged dotted path, per
prog.txt's REGOLA AGGIUNTIVA 2).

Part of #76.
