+++ dense_evolution/vhd_tb.py
@@ -1,187 +1,16 @@
+"""Backward-compatibility shim -- the real implementation moved to
+dense_evolution.solvers.vhd_tb as part of the Phase 2 subpackage split
+(see prog.txt). Kept so `from dense_evolution.vhd_tb import
+sp3s_star_hamiltonian` (used by external consumers, e.g.
+Dense-Evolution-Discovery) keeps working unchanged. Import from
+dense_evolution.solvers.vhd_tb directly in new code.
 """
-Vogl-Hjalmarson-Dow (1983) material-specific sp3s* tight-binding
-parameters -- a more accurate alternative to harrison_tb's universal
-(materials-independent) parameters, at the cost of needing a separately
-fitted parameter set per material.
+from dense_evolution.solvers.vhd_tb import (
+    Material, MATERIALS, sp3s_star_hamiltonian, direct_gap_at_gamma,
+    band_extrema_along_path,
+)
 
-Source: P. Vogl, H. P. Hjalmarson, J. D. Dow, "A Semi-empirical
-tight-binding theory of the electronic structure of semiconductors",
-J. Phys. Chem. Solids 44 (5), 365-378 (1983). Parameter values below are
-transcribed from an independent open-source implementation of the same
-table, github.com/rpmuller/TightBinding (`TB.py`, fetched 2026-08-05),
-which cites the same paper; not reproduced from the paper itself (not
-open access). Only materials appearing in that source are included.
-
-sp3s* basis: s, px, py, pz, and an extra excited s* orbital per atom
-(5 orbitals/atom, 10x10 Hamiltonian per zinc-blende/diamond 2-atom
-cell) -- the s* orbital exists purely to push the lowest conduction
-band down to the right energy without needing d-orbitals; it has no
-literal physical excited state behind it.
-
-Hermiticity note: the reference implementation's own H-builder forms
-the lower-left (cation-anion) block as `Hac.conjugate()` without
-transposing it, which does not in general produce a Hermitian matrix
-(the block mixes Vsapc/Vscpa asymmetrically). `_cation_anion_block`
-below is built the same way as that source's `get_Hac`, but assembled
-into the full Hamiltonian with a proper conjugate *transpose*
-(`.conj().T`) -- verified Hermitian for every material in `MATERIALS`
-(see the module's own smoke test).
-
-Validated against experiment (run 2026-08-05): GaAs direct gap at
-Gamma (k=0) computes to 1.55 eV vs. the experimental 1.42 eV (~9%
-high) -- compare to harrison_tb.zincblende_hamiltonian's 2.91 eV
-(~105% high) for the same material using Harrison's universal
-parameters. The improvement is expected: these parameters are fitted
-per material specifically to reproduce band-edge energies, unlike
-Harrison's one-table-fits-all approach.
-"""
-import numpy as np
-from collections import namedtuple
-
-__all__ = ['Material', 'MATERIALS', 'sp3s_star_hamiltonian', 'direct_gap_at_gamma',
-           'band_extrema_along_path']
-
-Material = namedtuple('Material', [
-    'name', 'Esa', 'Epa', 'Esc', 'Epc', 'Essa', 'Essc',
-    'Vss', 'Vxx', 'Vxy', 'Vsapc', 'Vscpa', 'Vssapc', 'Vsscpa',
-])
-
-# name: (Esa, Epa, Esc, Epc, Essa, Essc, Vss, Vxx, Vxy, Vsapc, Vscpa, Vssapc, Vsscpa)
-# 'a' = anion, 'c' = cation. All values in eV.
-MATERIALS = {
-    'C':    Material('C',    -4.5450, 3.8400, -4.5450, 3.8400, 11.3700, 11.3700,
-                      -22.7250, 3.8400, 11.6700, 15.2206, 15.2206, 8.2109, 8.2109),
-    'Si':   Material('Si',   -4.2000, 1.7150, -4.2000, 1.7150,  6.6850,  6.6850,
-                      -8.3000, 1.7150,  4.5750,  5.7292,  5.7292, 5.3749, 5.3749),
-    'Ge':   Material('Ge',   -5.8800, 1.6100, -5.8800, 1.6100,  6.3900,  6.3900,
-                      -6.7800, 1.6100,  4.9000,  5.4649,  5.4649, 5.2191, 5.2191),
-    'Sn':   Material('Sn',   -5.6700, 1.3300, -5.6700, 1.3300,  5.9000,  5.9000,
-                      -5.6700, 1.3300,  4.0800,  4.5116,  4.5116, 5.8939, 5.8939),
-    'SiC':  Material('SiC',  -8.4537, 2.1234, -4.8463, 4.3466,  9.6534,  9.3166,
-                      -12.4197, 3.0380, 5.9216,  9.4900,  9.2007, 8.7138, 4.4051),
-    'AlP':  Material('AlP',  -7.8466, 1.3169, -1.2534, 4.2831,  8.7069,  7.4231,
-                      -7.4535, 2.3749,  4.8378,  5.2451,  5.2775, 5.2508, 4.6388),
-    'AlAs': Material('AlAs', -7.5273, 0.9833, -1.1627, 3.5867,  7.4833,  6.7267,
-                      -6.6642, 1.8780,  4.2919,  5.1106,  5.4965, 4.5216, 4.9950),
-    'AlSb': Material('AlSb', -6.1714, 0.9807, -2.0716, 3.0163,  6.7607,  6.1543,
-                      -5.6448, 1.7199,  3.6648,  4.9121,  4.2137, 4.3662, 3.0739),
-    'GaP':  Material('GaP',  -8.1124, 1.1250, -2.1976, 4.1150,  8.5150,  7.1850,
-                      -7.4709, 2.1516,  5.1369,  4.2771,  6.3190, 4.6541, 5.0950),
-    'GaAs': Material('GaAs', -8.3431, 1.0414, -2.6569, 3.6686,  8.5914,  6.7386,
-                      -6.4513, 1.9546,  5.0779,  4.4800,  5.7839, 4.8422, 4.8077),
-    'GaSb': Material('GaSb', -7.3207, 0.8554, -3.8993, 2.9146,  6.6354,  5.9846,
-                      -6.1567, 1.5789,  4.1285,  4.9601,  4.6675, 4.9895, 4.2180),
-    'InP':  Material('InP',  -8.5274, 0.8735, -1.4826, 4.0465,  8.2635,  7.0665,
-                      -5.3614, 1.8801,  4.2324,  2.2265,  5.5825, 3.4623, 4.4814),
-    'InAs': Material('InAs', -9.5381, 0.9099, -2.7219, 3.7201,  7.4099,  6.7401,
-                      -5.6052, 1.8398,  4.4693,  3.0354,  5.4389, 3.3744, 3.9097),
-    'InSb': Material('InSb', -8.0157, 0.6738, -3.4643, 2.9162,  6.4530,  5.9362,
-                      -5.5193, 1.4018,  3.8761,  3.7880,  4.5900, 3.5666, 3.4048),
-    'ZnSe': Material('ZnSe', -11.8383, 1.5072, 0.0183, 5.9928,  7.5872,  8.9928,
-                      -6.2163, 3.0054,  5.9942,  3.4980,  6.3191, 2.5891, 3.9533),
-    'ZnTe': Material('ZnTe', -9.8150, 1.4834,  0.9350, 5.2666,  7.0834,  8.2666,
-                      -6.5765, 2.7951,  5.4670,  5.9827,  5.8199, 1.3196, 0.0000),
-}
-
-
-def _phase_factors(kxyz):
-    """g0..g3 structure factors for the 4 zinc-blende nearest-neighbor
-    bonds, for k given in units of 2*pi/a along the conventional cubic
-    axes (so Gamma=(0,0,0), X=(1,0,0), L=(0.5,0.5,0.5))."""
-    kxp, kyp, kzp = (np.pi / 2) * np.asarray(kxyz, dtype=float)
-    g0 = np.cos(kxp) * np.cos(kyp) * np.cos(kzp) - 1j * np.sin(kxp) * np.sin(kyp) * np.sin(kzp)
-    g1 = -np.cos(kxp) * np.sin(kyp) * np.sin(kzp) + 1j * np.sin(kxp) * np.cos(kyp) * np.cos(kzp)
-    g2 = -np.sin(kxp) * np.cos(kyp) * np.sin(kzp) + 1j * np.cos(kxp) * np.sin(kyp) * np.cos(kzp)
-    g3 = -np.sin(kxp) * np.sin(kyp) * np.cos(kzp) + 1j * np.cos(kxp) * np.cos(kyp) * np.sin(kzp)
-    return g0, g1, g2, g3
-
-
-def _cation_anion_block(mat, kxyz):
-    """5x5 anion-row x cation-column hopping block (s,px,py,pz,s* basis
-    on each side), before Hermitian assembly."""
-    g0, g1, g2, g3 = _phase_factors(kxyz)
-    return np.array([
-        [mat.Vss * g0,    mat.Vscpa * g1, mat.Vscpa * g2, mat.Vscpa * g3, 0],
-        [-mat.Vsapc * g1, mat.Vxx * g0,   mat.Vxy * g3,   mat.Vxy * g2,   -mat.Vssapc * g1],
-        [-mat.Vsapc * g2, mat.Vxy * g3,   mat.Vxx * g0,   mat.Vxy * g1,   -mat.Vssapc * g2],
-        [-mat.Vsapc * g3, mat.Vxy * g2,   mat.Vxy * g1,   mat.Vxx * g0,   -mat.Vssapc * g3],
-        [0,               mat.Vsscpa * g1, mat.Vsscpa * g2, mat.Vsscpa * g3, 0],
-    ], dtype=np.complex128)
-
-
-def sp3s_star_hamiltonian(kxyz, material):
-    """
-    10x10 sp3s* Bloch Hamiltonian, basis order
-    [anion: s,px,py,pz,s*, cation: s,px,py,pz,s*].
-
-    material: a Material namedtuple, or a key into MATERIALS.
-    kxyz: crystal momentum in units of 2*pi/a along the conventional
-    cubic axes (Gamma=(0,0,0), X=(1,0,0), L=(0.5,0.5,0.5)).
-    """
-    if isinstance(material, str):
-        if material not in MATERIALS:
-            raise ValueError(f"no VHD parameters for {material!r}; available: {sorted(MATERIALS)}")
-        material = MATERIALS[material]
-
-    Ha = np.diag([material.Esa, material.Epa, material.Epa, material.Epa, material.Essa])
-    Hc = np.diag([material.Esc, material.Epc, material.Epc, material.Epc, material.Essc])
-    Hac = _cation_anion_block(material, kxyz)
-
-    H = np.zeros((10, 10), dtype=np.complex128)
-    H[0:5, 0:5] = Ha
-    H[5:10, 5:10] = Hc
-    H[0:5, 5:10] = Hac
-    H[5:10, 0:5] = Hac.conj().T
-    return H
-
-
-def direct_gap_at_gamma(material):
-    """Conduction-band-minimum minus valence-band-maximum at k=Gamma,
-    for the 8 valence electrons (4 lowest of the 8 sp3-derived bands
-    filled; the 2 s* bands are always empty conduction states in this
-    model). Only meaningful as *the* fundamental gap for direct-gap
-    materials (e.g. GaAs) -- for indirect-gap materials (e.g. Si) this
-    is the Gamma-Gamma transition, not the true (lower) indirect gap,
-    which occurs off Gamma and this function does not compute."""
-    H = sp3s_star_hamiltonian((0.0, 0.0, 0.0), material)
-    eig = np.sort(np.linalg.eigvalsh(H).real)
-    return eig[4] - eig[3]
-
-
-def band_extrema_along_path(material, k_start, k_end, n_points=501):
-    """
-    Scans the valence-band maximum and conduction-band minimum (bands 3
-    and 4 of the 10, 0-indexed, sorted) along the straight line from
-    k_start to k_end (both in the same 2*pi/a cubic-axis units as
-    sp3s_star_hamiltonian), and returns
-    (vbm, vbm_k, cbm, cbm_k, gap) where vbm_k/cbm_k are the k-points
-    (same units) where each extremum was found.
-
-    Needed for indirect-gap materials (e.g. Si, Ge): the true
-    fundamental gap is not at Gamma, so direct_gap_at_gamma alone gives
-    the wrong (too-large) number for them. Example: Si's conduction
-    band minimum sits off-Gamma along Gamma->X (the Delta line), not
-    at Gamma itself.
-    """
-    if isinstance(material, str):
-        if material not in MATERIALS:
-            raise ValueError(f"no VHD parameters for {material!r}; available: {sorted(MATERIALS)}")
-        material = MATERIALS[material]
-
-    k_start = np.asarray(k_start, dtype=float)
-    k_end = np.asarray(k_end, dtype=float)
-    t = np.linspace(0.0, 1.0, n_points)
-    ks = k_start[None, :] + t[:, None] * (k_end - k_start)[None, :]
-
-    vbm_per_k = np.empty(n_points)
-    cbm_per_k = np.empty(n_points)
-    for i, k in enumerate(ks):
-        eig = np.sort(np.linalg.eigvalsh(sp3s_star_hamiltonian(k, material)).real)
-        vbm_per_k[i] = eig[3]
-        cbm_per_k[i] = eig[4]
-
-    vbm_idx = vbm_per_k.argmax()
-    cbm_idx = cbm_per_k.argmin()
-    vbm, vbm_k = vbm_per_k[vbm_idx], ks[vbm_idx]
-    cbm, cbm_k = cbm_per_k[cbm_idx], ks[cbm_idx]
-    return vbm, vbm_k, cbm, cbm_k, cbm - vbm
+__all__ = [
+    "Material", "MATERIALS", "sp3s_star_hamiltonian", "direct_gap_at_gamma",
+    "band_extrema_along_path",
+]
